requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
openai==0.27.8